@@ -3,6 +3,7 @@ module BenefitSponsors
   module Organizations
     module Factories
       class ProfileFactory
+        include BenefitSponsors::RegistrationHelper
         include ActiveModel::Validations
         include BenefitSponsors::Forms::NpnField
 
@@ -458,7 +459,7 @@ module BenefitSponsors
         def npn_already_taken?(npn)
           if is_broker_profile?
             if Person.where("broker_role.npn" => npn).any?
-              errors.add(:organization, "NPN has already been claimed by another broker. Please contact HBX-Customer Service - Call (855) 532-5465.")
+              errors.add(:organization, l10n("broker_agencies.profiles.npn_taken_error"))
               return true
             end
           end
