@@ -17,10 +17,10 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
     let(:current_effective_date)  { TimeKeeper.date_of_record.beginning_of_month }
     let(:site)                { create(:benefit_sponsors_site, :with_benefit_market, :as_hbx_profile, :cca) }
     let!(:benefit_market_catalog) { create(:benefit_markets_benefit_market_catalog, :with_product_packages,
-                                            benefit_market: benefit_market,
-                                            title: "SHOP Benefits for #{current_effective_date.year}",
-                                            application_period: (current_effective_date.beginning_of_year..current_effective_date.end_of_year))
-                                          }
+                                           benefit_market: benefit_market,
+                                           title: "SHOP Benefits for #{current_effective_date.year}",
+                                           application_period: (current_effective_date.beginning_of_year..current_effective_date.end_of_year))
+    }
     let(:benefit_market)      { site.benefit_markets.first }
     let!(:product_package) { benefit_market_catalog.product_packages.first }
 
@@ -32,15 +32,15 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
     let!(:employer_attestation)     { BenefitSponsors::Documents::EmployerAttestation.new(aasm_state: "approved") }
     let(:benefit_sponsorship) do
       FactoryGirl.create(
-        :benefit_sponsors_benefit_sponsorship,
-        :with_rating_area,
-        :with_service_areas,
-        supplied_rating_area: rating_area,
-        service_area_list: [service_area],
-        organization: organization,
-        profile_id: organization.profiles.first.id,
-        benefit_market: site.benefit_markets[0],
-        employer_attestation: employer_attestation)
+          :benefit_sponsors_benefit_sponsorship,
+          :with_rating_area,
+          :with_service_areas,
+          supplied_rating_area: rating_area,
+          service_area_list: [service_area],
+          organization: organization,
+          profile_id: organization.profiles.first.id,
+          benefit_market: site.benefit_markets[0],
+          employer_attestation: employer_attestation)
     end
 
     let(:start_on)  { current_effective_date }
@@ -56,66 +56,69 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
 
     let(:employee_role) { FactoryGirl.create(:benefit_sponsors_employee_role, person: person, employer_profile: benefit_sponsorship.profile, census_employee_id: census_employee.id) }
     let(:census_employee) { FactoryGirl.create(:census_employee,
-      employer_profile: benefit_sponsorship.profile,
-      benefit_sponsorship: benefit_sponsorship,
-      benefit_group_assignments: [benefit_group_assignment]
+                                               employer_profile: benefit_sponsorship.profile,
+                                               benefit_sponsorship: benefit_sponsorship,
+                                               benefit_group_assignments: [benefit_group_assignment]
     )}
     let(:person) { FactoryGirl.create(:person) }
     let(:family) { double }
     let(:hbx_enrollment) { double(kind: "employer_sponsored", effective_on: start_on, employee_role_id: employee_role.id,
-                            sponsored_benefit_package_id: benefit_package.id, benefit_group_assignment_id: benefit_group_assignment.id,
-                            aasm_state: 'coverage_selected') }
+                                  sponsored_benefit_package_id: benefit_package.id, benefit_group_assignment_id: benefit_group_assignment.id,
+                                  aasm_state: 'coverage_selected') }
 
-    before(:each) do
-      allow(ENV).to receive(:[]).with("fein").and_return(organization.fein)
+    around do |example|
+      ClimateControl.modify fein: organization.fein do
+        example.run
+      end
     end
 
     context "extend open enrollment" do
 
-        let!(:current_effective_date)  { TimeKeeper.date_of_record.next_month.beginning_of_month }
-        let!(:start_on)  { current_effective_date }
-        let!(:effective_period)  { start_on..start_on.next_year.prev_day }
-        let!(:ineligible_benefit_application) { FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, :with_benefit_package, benefit_sponsorship: benefit_sponsorship, aasm_state: "enrollment_ineligible", effective_period: effective_period)}
+      let!(:current_effective_date)  { TimeKeeper.date_of_record.next_month.beginning_of_month }
+      let!(:start_on)  { current_effective_date }
+      let!(:effective_period)  { start_on..start_on.next_year.prev_day }
+      let!(:ineligible_benefit_application) { FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, :with_benefit_package, benefit_sponsorship: benefit_sponsorship, aasm_state: "enrollment_ineligible", effective_period: effective_period)}
 
-        before do
-          allow(ENV).to receive(:[]).with("action").and_return("extend_open_enrollment")
-          allow(ENV).to receive(:[]).with("effective_date").and_return(start_on.strftime("%m/%d/%Y"))
-          allow(ENV).to receive(:[]).with("oe_end_date").and_return(start_on.prev_day.strftime("%m/%d/%Y"))
+      around do |example|
+        ClimateControl.modify action: 'extend_open_enrollment', effective_date: start_on.strftime("%m/%d/%Y"), oe_end_date: start_on.prev_day.strftime("%m/%d/%Y") do
+          example.run
         end
+      end
 
-        it "should extend open enrollment from enrollment ineligible" do
-          expect(ineligible_benefit_application.aasm_state).to eq :enrollment_ineligible
-          subject.migrate
-          ineligible_benefit_application.reload
-          benefit_sponsorship.reload
-          expect(ineligible_benefit_application.open_enrollment_period.max).to eq start_on.prev_day
-          expect(ineligible_benefit_application.aasm_state).to eq :enrollment_extended
-        end
+      it "should extend open enrollment from enrollment ineligible" do
+        expect(ineligible_benefit_application.aasm_state).to eq :enrollment_ineligible
+        subject.migrate
+        ineligible_benefit_application.reload
+        benefit_sponsorship.reload
+        expect(ineligible_benefit_application.open_enrollment_period.max).to eq start_on.prev_day
+        expect(ineligible_benefit_application.aasm_state).to eq :enrollment_extended
+      end
 
-        it "should extend open enrollment from enrollment open" do
-          ineligible_benefit_application.update_attributes!(aasm_state: "enrollment_open")
-          expect(ineligible_benefit_application.aasm_state).to eq :enrollment_open
-          subject.migrate
-          ineligible_benefit_application.reload
-          benefit_sponsorship.reload
-          expect(ineligible_benefit_application.open_enrollment_period.max).to eq start_on.prev_day
-          expect(ineligible_benefit_application.aasm_state).to eq :enrollment_extended
-        end
+      it "should extend open enrollment from enrollment open" do
+        ineligible_benefit_application.update_attributes!(aasm_state: "enrollment_open")
+        expect(ineligible_benefit_application.aasm_state).to eq :enrollment_open
+        subject.migrate
+        ineligible_benefit_application.reload
+        benefit_sponsorship.reload
+        expect(ineligible_benefit_application.open_enrollment_period.max).to eq start_on.prev_day
+        expect(ineligible_benefit_application.aasm_state).to eq :enrollment_extended
+      end
 
-        it "should not extend open enrollment from draft" do
-          ineligible_benefit_application.update_attributes!(aasm_state: "draft")
-          expect(ineligible_benefit_application.aasm_state).to eq :draft
-          expect { subject.migrate }.to raise_error("Unable to find benefit application!!")
-        end
+      it "should not extend open enrollment from draft" do
+        ineligible_benefit_application.update_attributes!(aasm_state: "draft")
+        expect(ineligible_benefit_application.aasm_state).to eq :draft
+        expect { subject.migrate }.to raise_error("Unable to find benefit application!!")
+      end
     end
 
     context "Update assm state to enrollment open" do
       context "update aasm state to enrollment open for non renewing ER" do
         let(:effective_date) { start_on }
 
-        before do
-          allow(ENV).to receive(:[]).with("action").and_return("begin_open_enrollment")
-          allow(ENV).to receive(:[]).with("effective_date").and_return(effective_date.strftime("%m/%d/%Y"))
+        around do |example|
+          ClimateControl.modify effective_date: effective_date.strftime("%m/%d/%Y"), action: 'begin_open_enrollment' do
+            example.run
+          end
         end
 
         it "should update the benefit application" do
@@ -139,9 +142,10 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
       context "update aasm state to enrollment open but not sponsoship for renewing ER" do
         let(:effective_date) { start_on }
 
-        before do
-          allow(ENV).to receive(:[]).with("action").and_return("begin_open_enrollment")
-          allow(ENV).to receive(:[]).with("effective_date").and_return(effective_date.strftime("%m/%d/%Y"))
+        around do |example|
+          ClimateControl.modify effective_date: effective_date.strftime("%m/%d/%Y"), action: 'begin_open_enrollment' do
+            example.run
+          end
         end
 
         it "should update the benefit application" do
@@ -158,52 +162,70 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
       end
     end
 
-    context "terminate benefit application", dbclean: :after_each do
-      let(:termination_date) { start_on.next_month.next_day }
-      let(:end_on)           { start_on.next_month.end_of_month }
+    context "terminate benefit application" do
+      let!(:termination_date) { start_on.next_month.next_day }
+      let!(:end_on)           { start_on.next_month.end_of_month }
 
-      before do
-        allow(ENV).to receive(:[]).with("termination_kind").and_return("voluntary")
-        allow(ENV).to receive(:[]).with("termination_reason").and_return("Company went out of business/bankrupt")
-        allow(ENV).to receive(:[]).with("notify_trading_partner").and_return("true")
-        allow(ENV).to receive(:[]).with('off_cycle_renewal').and_return('true')
-        allow(ENV).to receive(:[]).with("action").and_return("terminate")
-        allow(ENV).to receive(:[]).with("termination_date").and_return(termination_date.strftime("%m/%d/%Y"))
-        allow(ENV).to receive(:[]).with("end_on").and_return(end_on.strftime("%m/%d/%Y"))
-        subject.migrate
-        benefit_application.reload
+      around do |example|
+        ClimateControl.modify(
+            off_cycle_renewal: 'true',
+            termination_notice: 'true',
+            termination_kind: 'voluntary',
+            termination_reason: 'Company went out of business/bankrupt',
+            notify_trading_partner: 'true',
+            action: 'terminate',
+            termination_date: termination_date.strftime("%m/%d/%Y"),
+            end_on: end_on.strftime("%m/%d/%Y")
+        ) do
+          example.run
+        end
       end
 
       it "should terminate the benefit application" do
+        subject.migrate
+        benefit_application.reload
         expect(benefit_application.aasm_state).to eq :terminated
       end
 
       it "should transition benefit sponsorship to applicant" do
+        subject.migrate
         benefit_sponsorship.reload
         expect(benefit_sponsorship.aasm_state).to eq :applicant
       end
 
       it "should update end on date on benefit application" do
+        subject.migrate
+        benefit_application.reload
         expect(benefit_application.end_on).to eq end_on
       end
 
       it "should update end on date on benefit application" do
+        subject.migrate
+        benefit_application.reload
         expect(benefit_application.terminated_on).to eq termination_date
       end
 
       it "should update the termination kind" do
+        subject.migrate
+        benefit_application.reload
         expect(benefit_application.termination_kind).to eq "voluntary"
       end
 
       it "should update the termination reason" do
+        subject.migrate
+        benefit_application.reload
         expect(benefit_application.termination_reason).to eq "Company went out of business/bankrupt"
       end
 
       it "should terminate any active employee enrollments" do
+        subject.migrate
+        benefit_application.reload
         benefit_application.hbx_enrollments.each { |hbx_enrollment| expect(hbx_enrollment.aasm_state).to eq "coverage_terminated"}
       end
 
       it "should terminate any active employee enrollments with termination date as on Benefit Application" do
+        subject.migrate
+        benefit_application.reload
         benefit_application.hbx_enrollments.each { |hbx_enrollment| expect(hbx_enrollment.terminated_on).to eq end_on }
       end
     end
@@ -215,21 +237,36 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
       let!(:past_effective_period) {past_start_on..past_start_on.next_year.prev_day }
       let!(:mid_plan_year_effective_date) { Date.new(current_effective_date.year, 9, 1) }
       let!(:range_effective_period) { mid_plan_year_effective_date..mid_plan_year_effective_date.next_year.prev_day }
-      let!(:draft_benefit_application) { FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, :with_benefit_package, benefit_sponsorship: benefit_sponsorship, aasm_state: :imported, effective_period: past_effective_period)}
-      let!(:import_draft_benefit_application) { FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, :with_benefit_package, benefit_sponsorship: benefit_sponsorship, aasm_state: :imported, effective_period: range_effective_period)}
+      let(:draft_benefit_application) {
+        application = FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, benefit_sponsorship: benefit_sponsorship, aasm_state: :imported, effective_period: past_effective_period)
+        application.benefit_sponsor_catalog.save!
+        application.save
+        application
 
-      before :each do
-        allow(ENV).to receive(:[]).with("notify_trading_partner").and_return("true")
-        allow(ENV).to receive(:[]).with('action').and_return 'cancel'
-        allow(ENV).to receive(:[]).with('plan_year_start_on').and_return import_draft_benefit_application.effective_period.min.strftime("%m/%d/%Y")
-        subject.migrate
+      }
+
+      let!(:import_draft_benefit_application) {
+        application = FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, benefit_sponsorship: benefit_sponsorship, aasm_state: :imported, effective_period: range_effective_period)
+        application.benefit_sponsor_catalog.save!
+        application.save
+        application
+      }
+
+      around do |example|
+        ClimateControl.modify plan_year_start_on: range_effective_period.min.strftime("%m/%d/%Y"), action: 'cancel',notify_trading_partner: 'true' do
+          example.run
+        end
       end
 
       it "does not cancel non-imported draft benefit applications" do
+        benefit_sponsorship.reload
+        subject.migrate
         expect(draft_benefit_application.reload.aasm_state).to eq :imported
       end
 
       it "cancels import draft benefit applications" do
+        benefit_sponsorship.reload
+        subject.migrate
         expect(import_draft_benefit_application.reload.aasm_state).to eq :canceled
       end
     end
@@ -239,18 +276,22 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
       let(:new_start_date)   { effective_date.next_month }
       let(:new_end_date)     { new_start_date + 1.year }
       let(:effective_period) { effective_date..effective_date.next_year.prev_day }
-      
+
       let(:benefit_application) {
         application = FactoryGirl.create(:benefit_sponsors_benefit_application, :with_benefit_sponsor_catalog, benefit_sponsorship: benefit_sponsorship, effective_period: effective_period, aasm_state: :active)
         application.benefit_sponsor_catalog.save!
         application
       }
 
-      before do
-        allow(ENV).to receive(:[]).with("action").and_return("update_effective_period_and_approve")
-        allow(ENV).to receive(:[]).with("effective_date").and_return(effective_date.strftime("%m/%d/%Y"))
-        allow(ENV).to receive(:[]).with("new_start_date").and_return(new_start_date.strftime("%m/%d/%Y"))
-        allow(ENV).to receive(:[]).with("new_end_date").and_return(new_end_date.strftime("%m/%d/%Y"))
+      around do |example|
+        ClimateControl.modify(
+            action: 'update_effective_period_and_approve',
+            effective_date: effective_date.strftime("%m/%d/%Y"),
+            new_start_date: new_start_date.strftime("%m/%d/%Y"),
+            new_end_date: new_end_date.strftime("%m/%d/%Y")
+        ) do
+          example.run
+        end
       end
 
       it "should update the initial benefit application and transition the benefit sponsorship" do
@@ -273,10 +314,11 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
 
     context "Should force publish the benefit application", dbclean: :after_each do
       let(:effective_date) { start_on }
-      
-      before do
-        allow(ENV).to receive(:[]).with("action").and_return("force_submit_application")
-        allow(ENV).to receive(:[]).with("effective_date").and_return(effective_date.strftime("%m/%d/%Y"))
+
+      around do |example|
+        ClimateControl.modify effective_date: effective_date.strftime("%m/%d/%Y"), action: 'force_submit_application' do
+          example.run
+        end
       end
 
       it "should update the benefit application and transition the benefit sponsorship" do
@@ -356,11 +398,15 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
       let!(:old_benefit_package) { FactoryGirl.create(:benefit_sponsors_benefit_packages_benefit_package, benefit_application: old_benefit_application, product_package: product_package_1) }
       let!(:renewing_benefit_package) { FactoryGirl.create(:benefit_sponsors_benefit_packages_benefit_package, benefit_application: renewing_benefit_application, product_package: product_package_2) }
 
-      before do
-        allow(ENV).to receive(:[]).with("action").and_return("update_effective_period_and_approve")
-        allow(ENV).to receive(:[]).with("effective_date").and_return(effective_date.strftime("%m/%d/%Y"))
-        allow(ENV).to receive(:[]).with("new_start_date").and_return(new_start_date.strftime("%m/%d/%Y"))
-        allow(ENV).to receive(:[]).with("new_end_date").and_return(new_end_date.strftime("%m/%d/%Y"))
+      around do |example|
+        ClimateControl.modify(
+            action: 'update_effective_period_and_approve',
+            effective_date: effective_date.strftime("%m/%d/%Y"),
+            new_start_date: new_start_date.strftime("%m/%d/%Y"),
+            new_end_date: new_end_date.strftime("%m/%d/%Y")
+        ) do
+          example.run
+        end
       end
 
       it "should update the renewing benefit application and transition the benefit sponsorship" do
@@ -385,15 +431,19 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
       let(:termination_date) { start_on.next_month.next_day }
       let(:end_on)           { start_on.next_month.end_of_month }
 
-      before do
-        allow(ENV).to receive(:[]).with('off_cycle_renewal').and_return('true')
-        allow(ENV).to receive(:[]).with("termination_notice").and_return("true")
-        allow(ENV).to receive(:[]).with("termination_kind").and_return("voluntary")
-        allow(ENV).to receive(:[]).with("termination_reason").and_return("Company went out of business/bankrupt")
-        allow(ENV).to receive(:[]).with("notify_trading_partner").and_return("true")
-        allow(ENV).to receive(:[]).with("action").and_return("terminate")
-        allow(ENV).to receive(:[]).with("termination_date").and_return(termination_date.strftime("%m/%d/%Y"))
-        allow(ENV).to receive(:[]).with("end_on").and_return(end_on.strftime("%m/%d/%Y"))
+      around do |example|
+        ClimateControl.modify(
+            off_cycle_renewal: 'true',
+            termination_notice: 'true',
+            termination_kind: 'voluntary',
+            termination_reason: 'Company went out of business/bankrupt',
+            notify_trading_partner: 'true',
+            action: 'terminate',
+            termination_date: termination_date.strftime("%m/%d/%Y"),
+            end_on: end_on.strftime("%m/%d/%Y")
+        ) do
+          example.run
+        end
       end
 
       let(:model_instance) { benefit_application }
@@ -410,5 +460,5 @@ RSpec.describe ModifyBenefitApplication, dbclean: :after_each do
         end
       end
     end
-   end
+  end
 end
