@@ -498,12 +498,6 @@ RSpec.describe HbxEnrollment, type: :model, dbclean: :after_each do
         enrollment_for_waiver.propogate_waiver
         expect(benefit_group_assignment.aasm_state).not_to eq "coverage_waived"
       end
-
-      it "should cancel the shop enrollment" do
-        enrollment_for_waiver.propogate_waiver
-        existing_shop_enrollment.reload
-        expect(existing_shop_enrollment.aasm_state).to eq "coverage_canceled"
-      end
     end
   end
 
@@ -2739,6 +2733,7 @@ describe HbxEnrollment,"reinstate and change end date", type: :model, :dbclean =
       context "enrollment that already terminated with future date" do
         context "with new future termination date" do
           it "should update enrollment with new end date and notify enrollment" do
+            enrollment.update_attributes(terminated_on: TimeKeeper.date_of_record.next_month.end_of_month)
             expect(enrollment).to receive(:notify).with("acapi.info.events.hbx_enrollment.terminated", {:reply_to=>glue_event_queue_name, "hbx_enrollment_id" => enrollment.hbx_id, "enrollment_action_uri" => "urn:openhbx:terms:v1:enrollment#terminate_enrollment", "is_trading_partner_publishable" => false})
             enrollment.reterm_enrollment_with_earlier_date(TimeKeeper.date_of_record + 1.day, false)
             enrollment.reload
@@ -2779,6 +2774,7 @@ describe HbxEnrollment,"reinstate and change end date", type: :model, :dbclean =
       context "enrollment that already terminated with future date" do
         context "with new future termination date" do
           it "should update enrollment with new end date and notify enrollment" do
+            enrollment.update_attributes(terminated_on: TimeKeeper.date_of_record.next_month.end_of_month)
             expect(enrollment).to receive(:notify).with("acapi.info.events.hbx_enrollment.terminated", {:reply_to=>glue_event_queue_name, "hbx_enrollment_id" => enrollment.hbx_id, "enrollment_action_uri" => "urn:openhbx:terms:v1:enrollment#terminate_enrollment", "is_trading_partner_publishable" => false})
             enrollment.reterm_enrollment_with_earlier_date(TimeKeeper.date_of_record + 1.day, false)
             enrollment.reload
