@@ -90,7 +90,7 @@ describe Parsers::Xml::Cv::Importers::EnrollmentParser do
 
     it "should get applied_aptc_amount" do
       enrollment = subject.get_enrollment_object
-      expect(enrollment.applied_aptc_amount).to eq 194.65
+      expect(enrollment.applied_aptc_amount).to eq Money.new(19465)
     end
   end
 end
