@@ -27,7 +27,7 @@ RSpec.describe "_consumer_brokers_widget.html.erb" do
     end
 
     it "should display broker widget for consumer" do
-      expect(rendered).to have_selector('h3', "Your Broker")
+      expect(rendered).to have_text('h3', "Your Broker")
     end
 
     it "should display brokers email" do
@@ -60,15 +60,17 @@ RSpec.describe "_consumer_brokers_widget.html.erb" do
     end
 
     it "should display broker widget for consumer" do
-      expect(rendered).to have_selector('h3', "Select a Broker or Assister")
+      expect(rendered).to have_text('h3', "Select a Broker or Assister")
     end
 
     it "should display get help signing up button" do
-      expect(rendered).to have_selector('a', "Get Help Signing Up")
+      expect(rendered).to have_text('a', "Get Help Signing Up")
     end
 
     it "should display get help signing up button" do
-      expect(rendered).to have_selector('a', "Find Assistance Another Way")
+      # I know I'm flipping this test but when it was previously have_selector it wasn't working correctly
+      # when broker_agency.present? returns false this should not be rendered, check the view, line 55 # from DC.
+      expect(rendered).to_not have_text('a', 'Find Assistance Another Way')
     end
 
   end
