@@ -34,8 +34,7 @@ RSpec.describe "insured/families/_employer_panel.html.erb" do
 
     it "should have employer name" do
       expect(rendered).to have_selector('input')
-      expect(rendered).to have_content("Congratulations on your new job at 
-                                        #{person.employee_roles.first.employer_profile.legal_name}.")
+      expect(rendered).to have_content("Congratulations on your new job at #{person.employee_roles.first.employer_profile.legal_name}.")
     end
   end
 
@@ -64,10 +63,8 @@ RSpec.describe "insured/families/_employer_panel.html.erb" do
 
     it "should have notices for both employers" do
       expect(rendered).to have_selector('input')
-      expect(rendered).to have_content("Congratulations on your new job at 
-                                        #{person.employee_roles[0].employer_profile.legal_name}.")
-      expect(rendered).to have_content("Congratulations on your new job at 
-                                        #{person.employee_roles[1].employer_profile.legal_name}.")
+      expect(rendered).to have_content("Congratulations on your new job at #{person.employee_roles[0].employer_profile.legal_name}.")
+      expect(rendered).to have_content("Congratulations on your new job at #{person.employee_roles[1].employer_profile.legal_name}.")
     end
   end
 end
