@@ -4,13 +4,13 @@ RSpec.describe "insured/families/_navigation.html.erb" do
   let(:person) {FactoryGirl.create(:person, :with_family)}
   let(:user){ FactoryGirl.create(:user, person: person) }
 
-  let(:person){
-    instance_double(
-      "Person",
-      first_name: "my first name",
-      last_name: "my last name"
-      )
-  }
+  # let(:person){
+  #   instance_double(
+  #     "Person",
+  #     first_name: "my first name",
+  #     last_name: "my last name"
+  #     )
+  # }
   let(:employee_role){ instance_double("EmployeeRole") }
   let(:employer_profile){ instance_double("EmployerProfile") }
   let(:broker_agency_profile){ instance_double("BrokerAgencyProfile") }
@@ -62,9 +62,9 @@ RSpec.describe "insured/families/_navigation.html.erb" do
       end
 
       it "should display the info under my account in the side bar" do
-        expect(rendered).to have_selector('h4', 'My Account')
-        expect(rendered).to have_selector('span', "#{person.first_name} #{person.last_name}")
-        expect(rendered).to have_selector('span', "Your household of #{family_members.count}")
+        expect(rendered).to have_text('h4', 'My Account')
+        expect(rendered).to have_text('span', "#{person.first_name} #{person.last_name}")
+        expect(rendered).to have_text('span', "Your Household of #{family_members.count}")
       end
 
       it "should display the list of active household" do
@@ -96,9 +96,9 @@ RSpec.describe "insured/families/_navigation.html.erb" do
       end
 
       it "should display the info under my account in the side bar" do
-        expect(rendered).to have_selector('h4', 'My Account')
-        expect(rendered).to have_selector('span', "#{person.first_name} #{person.last_name}")
-        expect(rendered).to have_selector('span', "Your household of #{family_members.count}")
+        expect(rendered).to have_text( 'My Account')
+        expect(rendered).to have_text('span', "#{person.first_name} #{person.last_name}")
+        expect(rendered).to have_text("Your Household of #{family_members.count}")
       end
 
       it "should display the list of active household" do
@@ -132,9 +132,9 @@ RSpec.describe "insured/families/_navigation.html.erb" do
       end
 
       it "should display the info under my account in the side bar" do
-        expect(rendered).to have_selector('h4', 'My Account')
-        expect(rendered).to have_selector('span', "#{person.first_name} #{person.last_name}")
-        expect(rendered).to have_selector('span', "Your household of #{family_members.count}")
+        expect(rendered).to have_text('h4', 'My Account')
+        expect(rendered).to have_text('span', "#{person.first_name} #{person.last_name}")
+        expect(rendered).to have_text('span', "Your Household of #{family_members.count}")
       end
 
       it "should display the list of active household" do
@@ -168,9 +168,9 @@ RSpec.describe "insured/families/_navigation.html.erb" do
     end
 
     it "should display the info under my account in the side bar" do
-      expect(rendered).to have_selector('h4', 'My Account')
-      expect(rendered).to have_selector('span', "#{person.first_name} #{person.last_name}")
-      expect(rendered).to have_selector('span', "Your household of #{family_members.count}")
+      expect(rendered).to have_text('h4', 'My Account')
+      expect(rendered).to have_text('span', "#{person.first_name} #{person.last_name}")
+      expect(rendered).to have_text('span', "Your Household of #{family_members.count}")
     end
 
     it "should display the list of active household" do
